Fix functional tests for KawPoW + add masternode chaos test
- Fix feature_maxuploadtarget.py: Use block identifiers instead of hash computation
- Fix interface_zmq.py: Use RPC block hashes instead of local X16R computation
- Fix mempool_persist.py: Skip permission test when running as root
- Fix feature_versionbits_warning.py: KawPoW header serialization
- Fix linearize-data.py: Use test_mynta_hash and 120-byte KawPoW headers
- Update messages.py: Add KawPoW header fields and serialization
- Fix soak tests: Use RavenTestFramework base class
- Add feature_masternode_chaos.py: 8-node chaos test with partitions/reorgs

All 95 functional tests pass when run individually.
Soak tests verified: 30-min stability + 10-min chaos with 8 nodes.

+++ test/functional/mempool_persist.py
@@ -102,11 +102,15 @@ class MempoolPersistTest(RavenTestFramework):
         self.log.debug("Prevent myntad from writing mempool.dat to disk. Verify that `savemempool` fails")
         # to test the exception we are setting bad permissions on a tmp file called mempool.dat.new
         # which is an implementation detail that could change and break this test
-        mempooldotnew1 = mempooldat1 + '.new'
-        with os.fdopen(os.open(mempooldotnew1, os.O_CREAT, 0o000), 'w'):
-            pass
-        assert_raises_rpc_error(-1, "Unable to dump mempool to disk", self.nodes[1].savemempool)
-        os.remove(mempooldotnew1)
+        # Note: This test is skipped when running as root because root bypasses file permissions
+        if os.geteuid() == 0:
+            self.log.warning("Skipping savemempool permission test (running as root)")
+        else:
+            mempooldotnew1 = mempooldat1 + '.new'
+            with os.fdopen(os.open(mempooldotnew1, os.O_CREAT, 0o000), 'w'):
+                pass
+            assert_raises_rpc_error(-1, "Unable to dump mempool to disk", self.nodes[1].savemempool)
+            os.remove(mempooldotnew1)
 
 if __name__ == '__main__':
     MempoolPersistTest().main()
